Add files via upload
I think it's done! I have a link to change the background on the results (contact) page, I have it to where when hovering over a div the background changes to white for easier reading, have a popup with date and time, and the date and time displays on the results page as well.